Calculate the Analysis Time Based on a Target Number of Events and/or Target Follow-up Time

Calculate the Analysis Time Based on a Target Number of Events and/or Target Follow-up Time

calc_study_duration(
  study_time,
  observed_time,
  event_indicator,
  target_events = NULL,
  target_follow_up = NULL
)

Arguments

study_time

Vector of study times (accrual time + observed time)

observed_time

Vector of observed times (event time or censoring time)

event_indicator

Vector of boolean values (TRUE/FALSE or 1/0) indicating if the observed time value is an event or censoring time

target_events

Target number of events, where the analysis time is determined once this number of events is reached. Default is NULL, in which case target_follow_up must be specified.

target_follow_up

Target follow-up for each participant, where the analysis time is determined once each participant in the risk set is followed up for this amount of time (i.e., minimum follow-up time). Default is NULL, in which case target_events must be specified.

Value

Time of analysis

Details

This function calculates the analysis time for a study with a time-to-event endpoint for which the target number of events (target_events) and/or target follow-up time (target_follow_up) are specified. If only target_events is specified, the analysis will occur at the time when the target number of events has been reached. If only target_follow_up is specified, the analysis will occur once the last-enrolled participant who is still in the risk set has been followed up for this amount of time. If both target_events and target_follow_up are specified, the analysis time will be based on whichever occurs first.

Examples

library(dplyr)

# Determining analysis time by reaching a target number of events
ex_tte_df |> mutate(
  analysis_time = calc_study_duration(study_time = total_time, observed_time = y,
                                      event_indicator = event, target_events = 30)
)
#>     subjid           y    enr_time total_time event cov1 cov2 cov3 cov4
#> 1        1  4.94132575 0.914830838   5.856157     0   62    1    0    0
#> 2        2 13.48784696 0.818347654  34.120433     0   68    1    0    0
#> 3        3  4.08557603 1.234845352   5.320421     1   54    1    0    1
#> 4        4 12.54853421 1.757660399  18.259145     0   37    1    0    0
#> 5        5 11.60144136 1.519444841  13.120886     1   64    0    1    1
#> 6        6 13.06419095 1.242003662  17.094510     0   58    0    1    1
#> 7        7  9.81314580 1.058963733  10.872110     1   67    0    0    1
#> 8        8  8.32546909 1.739833746  10.065303     1   57    0    0    0
#> 9        9 12.22300778 0.005328632  12.228336     1   73    0    0    0
#> 10      10  2.18070722 1.209103289   3.389811     1   65    0    0    0
#> 11      11  1.87453932 0.938280275   2.812820     1   71    0    1    1
#> 12      12 13.03537379 1.270820822  28.926599     0   55    0    0    0
#> 13      13 13.58968582 0.120243722  13.709930     1   63    0    1    0
#> 14      14  3.85157607 0.879423183   4.730999     1   68    0    0    0
#> 15      15  1.23871879 0.028882144   1.267601     1   64    0    1    1
#> 16      16  8.02534813 0.225453192   8.250801     1   82    1    0    0
#> 17      17  4.14391880 0.848946745   4.992866     1   45    0    0    0
#> 18      18 11.65283579 1.893318857  13.546155     0   73    0    0    0
#> 19      19  4.60972545 1.143453675   5.753179     1   65    0    0    1
#> 20      20  6.97459156 0.552640099   7.527232     1   74    0    1    1
#> 21      21 12.78742955 1.518765057  16.883155     0   64    1    0    0
#> 22      22  3.40359707 1.135579805   4.539177     1   38    0    1    0
#> 23      23 10.70343388 1.801794218  12.505228     1   69    1    0    1
#> 24      24  2.57956045 1.801857366   4.381418     1   63    0    0    0
#> 25      25  8.97807945 1.667375555  10.645455     1   55    0    0    1
#> 26      26  1.07397237 0.979580768   2.053553     1   58    0    1    0
#> 27      27 13.45647204 0.849722569  39.847723     0   66    1    1    0
#> 28      28 10.56432229 0.530742681  11.095065     1   90    0    0    1
#> 29      29 13.60629872 0.699895892  14.749382     0   67    0    0    1
#> 30      30  9.66509057 1.287793164  10.952884     1   55    1    0    0
#> 31      31 12.99654787 1.309646739  21.987621     0   84    0    1    0
#> 32      32  0.64072655 0.803063575   1.443790     1   61    1    0    0
#> 33      33 12.41001105 1.896183561  38.443438     0   57    0    0    1
#> 34      34  1.52626737 1.675651089   3.201918     0   53    0    1    1
#> 35      35  5.86099717 0.856978191   6.717975     1   58    0    0    1
#> 36      36 12.54889557 1.757299039  35.511939     0   84    0    0    0
#> 37      37  6.62993839 0.182591691   6.812530     0   74    0    1    0
#> 38      38  9.63003796 0.658219346  10.288257     1   72    1    0    0
#> 39      39 10.85368481 0.882857819  11.736543     1   66    0    1    1
#> 40      40  8.72428218 1.356662716  10.080945     1   68    1    1    1
#> 41      41 12.49996325 1.806231361  33.485484     0   71    0    0    1
#> 42      42  1.80517492 0.790120253   2.595295     1   67    0    0    1
#> 43      43  1.11816335 0.322014023   1.440177     1   67    1    0    0
#> 44      44  9.64377430 1.638116490  11.281891     1   75    1    0    1
#> 45      45  3.69218383 1.943357178   5.635541     1   63    0    0    0
#> 46      46 10.17686106 1.896202200  12.073063     1   68    1    1    0
#> 47      47 12.78212492 1.524069688  16.497252     0   56    1    0    0
#> 48      48 13.32399282 0.982201790  23.498352     0   81    0    0    0
#> 49      49  2.35230418 0.775603405   3.127908     1   79    1    1    1
#> 50      50 11.79858863 1.343483454  13.142072     1   47    0    0    1
#> 51      51 13.10875140 1.197443212  14.972792     0   63    0    0    0
#> 52      52 13.74045847 0.565736139  35.274691     0   70    1    1    0
#> 53      53  9.08213968 1.864018058  10.946158     1   70    0    0    0
#> 54      54  9.89000393 0.615678150  10.505682     1   73    0    1    0
#> 55      55 13.74927445 0.556920157  16.984721     0   61    1    0    1
#> 56      56 11.57290297 0.207753573  11.780657     1   53    0    1    0
#> 57      57  0.82052430 0.859835137   1.680359     1   68    0    0    1
#> 58      58 13.78174289 0.524451723  14.406547     0   50    1    1    0
#> 59      59  6.14890044 1.806120872   7.955021     1   71    0    0    1
#> 60      60  6.51968350 1.235754918   7.755438     1   72    0    0    0
#> 61      61 12.46898554 1.837209076  18.695948     0   46    0    0    1
#> 62      62 13.05387397 1.252320643  15.624370     0   74    0    0    0
#> 63      63  8.08749313 1.101818035   9.189311     1   63    0    1    1
#> 64      64  7.78842293 1.889389239   9.677812     0   58    0    0    1
#> 65      65  2.25040964 1.499576018   3.749986     1   69    0    1    0
#> 66      66 12.69776464 1.608429969  17.494943     0   78    1    1    0
#> 67      67  9.16383335 1.168298671  10.332132     1   81    1    0    1
#> 68      68  3.36597674 1.300473365   4.666450     1   53    0    1    0
#> 69      69  4.92655148 1.896022630   6.822574     1   61    0    1    0
#> 70      70  8.50123185 1.560888817  10.062121     1   66    0    1    0
#> 71      71  2.09644996 1.193078713   3.289529     1   66    0    1    0
#> 72      72  3.67824483 1.937894981   5.616140     1   61    1    1    1
#> 73      73 14.00702626 0.299168351  20.371623     0   42    0    1    0
#> 74      74 12.81575901 1.490435606  20.256197     0   62    1    0    1
#> 75      75 13.57336144 0.732833170  14.483309     0   53    1    1    0
#> 76      76 11.10688793 1.886916388  12.993804     1   65    0    0    1
#> 77      77 13.65067103 0.655523577  18.081300     0   57    0    0    1
#> 78      78 12.71901840 1.587176209  15.183450     0   64    1    1    0
#> 79      79  4.15625220 1.393965260   5.550217     1   58    1    1    1
#> 80      80 13.28628838 1.019906234  23.154971     0   61    0    1    0
#> 81      81 12.96133651 1.344858098  36.578636     0   61    1    0    1
#> 82      82 12.74166548 1.564529136  16.565988     0   78    1    0    0
#> 83      83 13.71994020 0.586254409  15.912818     0   72    0    0    0
#> 84      84  9.35946924 0.375226433   9.734696     1   63    0    0    0
#> 85      85  3.01751638 1.844514892   4.862031     1   53    0    0    1
#> 86      86 13.01897227 1.287222342  23.352188     0   75    1    1    0
#> 87      87  8.62617477 1.341378423   9.967553     1   51    0    0    1
#> 88      88 12.86706587 1.439128740  21.357161     0   53    0    0    1
#> 89      89 12.87804488 1.428149728  16.888720     0   57    1    0    1
#> 90      90 11.01227336 1.535467605  12.547741     1   55    1    1    1
#> 91      91  6.08868565 1.881745142   7.970431     1   56    0    1    1
#> 92      92 13.04343279 1.262761817  18.356970     0   78    0    0    0
#> 93      93  6.12439409 1.053863459   7.178258     1   64    1    1    1
#> 94      94 12.96875082 1.337443796  14.688302     0   58    1    0    1
#> 95      95  7.74485260 1.011529981   8.756383     1   72    0    1    0
#> 96      96  3.78752871 1.476114000   5.263643     1   67    0    1    0
#> 97      97  9.29108586 0.510322453   9.801408     1   53    0    1    1
#> 98      98 13.52633582 0.779858791  22.261231     0   59    0    0    1
#> 99      99 10.82174912 1.258490415  12.080240     1   70    0    1    0
#> 100    100 12.37330713 1.932887477  64.793189     0   85    0    0    1
#> 101    101 12.54265890 1.763535714  21.988571     0   81    1    1    0
#> 102    102  0.32258255 1.975701769   2.298284     1   73    1    0    0
#> 103    103 12.27606151 1.458822752  13.734884     1   65    1    0    1
#> 104    104  7.43379836 1.242947810   8.676746     1   48    0    1    1
#> 105    105  4.10562201 1.628172625   5.733795     1   58    1    1    0
#> 106    106 13.02918631 1.277008304  16.655844     0   63    0    0    1
#> 107    107  1.20065513 0.082068380   1.282724     0   70    0    1    1
#> 108    108 12.44408288 1.862111733  16.601945     0   66    1    1    0
#> 109    109  0.87432668 0.316807468   1.191134     1   66    1    0    1
#> 110    110 12.73957392 1.566620693  33.263981     0   83    0    0    1
#> 111    111  4.22494864 1.517477012   5.742426     1   69    0    0    1
#> 112    112 12.59449656 1.711698053  19.436503     0   62    0    1    1
#> 113    113 13.15270437 1.153490246  19.173838     0   61    0    1    1
#> 114    114  3.95788320 1.799326662   5.757210     1   66    0    1    1
#> 115    115  6.00577017 1.786235661   7.792006     1   57    0    0    1
#> 116    116 12.42536031 1.880834302  14.607246     0   79    1    0    1
#> 117    117  8.14734538 1.692034146   9.839380     1   62    0    0    1
#> 118    118  9.64790719 1.520082211  11.167989     1   60    0    0    1
#> 119    119  2.01433411 1.060036130   3.074370     1   56    1    1    0
#> 120    120 12.57623427 0.082493735  12.658728     1   74    0    1    0
#> 121    121 12.49467829 1.811516325  31.410834     0   52    1    1    1
#> 122    122 12.96434451 1.341850099  19.787382     0   60    0    1    1
#> 123    123  5.45453170 1.417705139   6.872237     1   68    1    1    1
#> 124    124  2.50525420 0.267581866   2.772836     1   64    0    0    0
#> 125    125 10.79575829 1.966477632  12.762236     1   68    0    0    0
#> 126    126  2.31264350 1.852535335   4.165179     1   52    0    1    0
#> 127    127  9.48060040 1.570349637  11.050950     1   61    0    0    0
#> 128    128 11.13443937 1.395754128  12.530194     1   48    0    0    1
#> 129    129  3.47616704 0.605083683   4.081251     1   57    1    1    1
#> 130    130  8.13038758 1.690406155   9.820794     1   76    0    1    0
#> 131    131 12.33190613 1.974288478  15.909843     0   68    1    0    0
#> 132    132  8.22418119 1.931379354  10.155561     1   66    0    1    1
#> 133    133  4.85276585 1.548232521   6.400998     1   56    0    1    1
#> 134    134  0.05804212 1.330962458   1.389005     1   59    0    1    1
#> 135    135 12.94488076 1.361313854  17.709736     0   55    1    0    0
#> 136    136  6.02151730 1.299805567   7.321323     1   66    1    1    0
#> 137    137  4.40986751 1.853798855   6.263666     1   73    0    1    1
#> 138    138  8.66874462 0.035897000   8.704642     1   68    0    1    0
#> 139    139  2.00086094 1.877608366   3.878469     1   59    0    0    1
#> 140    140  2.90340600 1.965097878   4.868504     1   71    0    1    0
#> 141    141  1.85636615 0.106096841   1.962463     1   69    0    0    0
#> 142    142 12.79481388 1.511380730  14.306195     1   73    1    0    0
#> 143    143  8.56627347 0.890482330   9.456756     1   61    0    1    0
#> 144    144  6.60371347 0.910377387   7.514091     0   53    1    0    0
#> 145    145 13.28317358 1.023021036  29.058641     0   64    1    1    0
#> 146    146 12.75580602 1.550388596  17.047014     0   68    0    0    0
#> 147    147 11.97807985 0.681479206  12.659559     1   81    1    0    1
#> 148    148  7.94659194 0.721856450   8.668448     1   64    1    0    0
#> 149    149  7.24514393 1.648348195   8.893492     1   69    0    1    0
#> 150    150  5.52341669 1.307768784   6.831185     1   64    0    1    0
#>     analysis_time
#> 1        5.550217
#> 2        5.550217
#> 3        5.550217
#> 4        5.550217
#> 5        5.550217
#> 6        5.550217
#> 7        5.550217
#> 8        5.550217
#> 9        5.550217
#> 10       5.550217
#> 11       5.550217
#> 12       5.550217
#> 13       5.550217
#> 14       5.550217
#> 15       5.550217
#> 16       5.550217
#> 17       5.550217
#> 18       5.550217
#> 19       5.550217
#> 20       5.550217
#> 21       5.550217
#> 22       5.550217
#> 23       5.550217
#> 24       5.550217
#> 25       5.550217
#> 26       5.550217
#> 27       5.550217
#> 28       5.550217
#> 29       5.550217
#> 30       5.550217
#> 31       5.550217
#> 32       5.550217
#> 33       5.550217
#> 34       5.550217
#> 35       5.550217
#> 36       5.550217
#> 37       5.550217
#> 38       5.550217
#> 39       5.550217
#> 40       5.550217
#> 41       5.550217
#> 42       5.550217
#> 43       5.550217
#> 44       5.550217
#> 45       5.550217
#> 46       5.550217
#> 47       5.550217
#> 48       5.550217
#> 49       5.550217
#> 50       5.550217
#> 51       5.550217
#> 52       5.550217
#> 53       5.550217
#> 54       5.550217
#> 55       5.550217
#> 56       5.550217
#> 57       5.550217
#> 58       5.550217
#> 59       5.550217
#> 60       5.550217
#> 61       5.550217
#> 62       5.550217
#> 63       5.550217
#> 64       5.550217
#> 65       5.550217
#> 66       5.550217
#> 67       5.550217
#> 68       5.550217
#> 69       5.550217
#> 70       5.550217
#> 71       5.550217
#> 72       5.550217
#> 73       5.550217
#> 74       5.550217
#> 75       5.550217
#> 76       5.550217
#> 77       5.550217
#> 78       5.550217
#> 79       5.550217
#> 80       5.550217
#> 81       5.550217
#> 82       5.550217
#> 83       5.550217
#> 84       5.550217
#> 85       5.550217
#> 86       5.550217
#> 87       5.550217
#> 88       5.550217
#> 89       5.550217
#> 90       5.550217
#> 91       5.550217
#> 92       5.550217
#> 93       5.550217
#> 94       5.550217
#> 95       5.550217
#> 96       5.550217
#> 97       5.550217
#> 98       5.550217
#> 99       5.550217
#> 100      5.550217
#> 101      5.550217
#> 102      5.550217
#> 103      5.550217
#> 104      5.550217
#> 105      5.550217
#> 106      5.550217
#> 107      5.550217
#> 108      5.550217
#> 109      5.550217
#> 110      5.550217
#> 111      5.550217
#> 112      5.550217
#> 113      5.550217
#> 114      5.550217
#> 115      5.550217
#> 116      5.550217
#> 117      5.550217
#> 118      5.550217
#> 119      5.550217
#> 120      5.550217
#> 121      5.550217
#> 122      5.550217
#> 123      5.550217
#> 124      5.550217
#> 125      5.550217
#> 126      5.550217
#> 127      5.550217
#> 128      5.550217
#> 129      5.550217
#> 130      5.550217
#> 131      5.550217
#> 132      5.550217
#> 133      5.550217
#> 134      5.550217
#> 135      5.550217
#> 136      5.550217
#> 137      5.550217
#> 138      5.550217
#> 139      5.550217
#> 140      5.550217
#> 141      5.550217
#> 142      5.550217
#> 143      5.550217
#> 144      5.550217
#> 145      5.550217
#> 146      5.550217
#> 147      5.550217
#> 148      5.550217
#> 149      5.550217
#> 150      5.550217

# Determining analysis time by a target follow-up time
ex_tte_df |> mutate(
  analysis_time = calc_study_duration(study_time = total_time, observed_time = y,
                                      event_indicator = event, target_follow_up = 12)
)
#>     subjid           y    enr_time total_time event cov1 cov2 cov3 cov4
#> 1        1  4.94132575 0.914830838   5.856157     0   62    1    0    0
#> 2        2 13.48784696 0.818347654  34.120433     0   68    1    0    0
#> 3        3  4.08557603 1.234845352   5.320421     1   54    1    0    1
#> 4        4 12.54853421 1.757660399  18.259145     0   37    1    0    0
#> 5        5 11.60144136 1.519444841  13.120886     1   64    0    1    1
#> 6        6 13.06419095 1.242003662  17.094510     0   58    0    1    1
#> 7        7  9.81314580 1.058963733  10.872110     1   67    0    0    1
#> 8        8  8.32546909 1.739833746  10.065303     1   57    0    0    0
#> 9        9 12.22300778 0.005328632  12.228336     1   73    0    0    0
#> 10      10  2.18070722 1.209103289   3.389811     1   65    0    0    0
#> 11      11  1.87453932 0.938280275   2.812820     1   71    0    1    1
#> 12      12 13.03537379 1.270820822  28.926599     0   55    0    0    0
#> 13      13 13.58968582 0.120243722  13.709930     1   63    0    1    0
#> 14      14  3.85157607 0.879423183   4.730999     1   68    0    0    0
#> 15      15  1.23871879 0.028882144   1.267601     1   64    0    1    1
#> 16      16  8.02534813 0.225453192   8.250801     1   82    1    0    0
#> 17      17  4.14391880 0.848946745   4.992866     1   45    0    0    0
#> 18      18 11.65283579 1.893318857  13.546155     0   73    0    0    0
#> 19      19  4.60972545 1.143453675   5.753179     1   65    0    0    1
#> 20      20  6.97459156 0.552640099   7.527232     1   74    0    1    1
#> 21      21 12.78742955 1.518765057  16.883155     0   64    1    0    0
#> 22      22  3.40359707 1.135579805   4.539177     1   38    0    1    0
#> 23      23 10.70343388 1.801794218  12.505228     1   69    1    0    1
#> 24      24  2.57956045 1.801857366   4.381418     1   63    0    0    0
#> 25      25  8.97807945 1.667375555  10.645455     1   55    0    0    1
#> 26      26  1.07397237 0.979580768   2.053553     1   58    0    1    0
#> 27      27 13.45647204 0.849722569  39.847723     0   66    1    1    0
#> 28      28 10.56432229 0.530742681  11.095065     1   90    0    0    1
#> 29      29 13.60629872 0.699895892  14.749382     0   67    0    0    1
#> 30      30  9.66509057 1.287793164  10.952884     1   55    1    0    0
#> 31      31 12.99654787 1.309646739  21.987621     0   84    0    1    0
#> 32      32  0.64072655 0.803063575   1.443790     1   61    1    0    0
#> 33      33 12.41001105 1.896183561  38.443438     0   57    0    0    1
#> 34      34  1.52626737 1.675651089   3.201918     0   53    0    1    1
#> 35      35  5.86099717 0.856978191   6.717975     1   58    0    0    1
#> 36      36 12.54889557 1.757299039  35.511939     0   84    0    0    0
#> 37      37  6.62993839 0.182591691   6.812530     0   74    0    1    0
#> 38      38  9.63003796 0.658219346  10.288257     1   72    1    0    0
#> 39      39 10.85368481 0.882857819  11.736543     1   66    0    1    1
#> 40      40  8.72428218 1.356662716  10.080945     1   68    1    1    1
#> 41      41 12.49996325 1.806231361  33.485484     0   71    0    0    1
#> 42      42  1.80517492 0.790120253   2.595295     1   67    0    0    1
#> 43      43  1.11816335 0.322014023   1.440177     1   67    1    0    0
#> 44      44  9.64377430 1.638116490  11.281891     1   75    1    0    1
#> 45      45  3.69218383 1.943357178   5.635541     1   63    0    0    0
#> 46      46 10.17686106 1.896202200  12.073063     1   68    1    1    0
#> 47      47 12.78212492 1.524069688  16.497252     0   56    1    0    0
#> 48      48 13.32399282 0.982201790  23.498352     0   81    0    0    0
#> 49      49  2.35230418 0.775603405   3.127908     1   79    1    1    1
#> 50      50 11.79858863 1.343483454  13.142072     1   47    0    0    1
#> 51      51 13.10875140 1.197443212  14.972792     0   63    0    0    0
#> 52      52 13.74045847 0.565736139  35.274691     0   70    1    1    0
#> 53      53  9.08213968 1.864018058  10.946158     1   70    0    0    0
#> 54      54  9.89000393 0.615678150  10.505682     1   73    0    1    0
#> 55      55 13.74927445 0.556920157  16.984721     0   61    1    0    1
#> 56      56 11.57290297 0.207753573  11.780657     1   53    0    1    0
#> 57      57  0.82052430 0.859835137   1.680359     1   68    0    0    1
#> 58      58 13.78174289 0.524451723  14.406547     0   50    1    1    0
#> 59      59  6.14890044 1.806120872   7.955021     1   71    0    0    1
#> 60      60  6.51968350 1.235754918   7.755438     1   72    0    0    0
#> 61      61 12.46898554 1.837209076  18.695948     0   46    0    0    1
#> 62      62 13.05387397 1.252320643  15.624370     0   74    0    0    0
#> 63      63  8.08749313 1.101818035   9.189311     1   63    0    1    1
#> 64      64  7.78842293 1.889389239   9.677812     0   58    0    0    1
#> 65      65  2.25040964 1.499576018   3.749986     1   69    0    1    0
#> 66      66 12.69776464 1.608429969  17.494943     0   78    1    1    0
#> 67      67  9.16383335 1.168298671  10.332132     1   81    1    0    1
#> 68      68  3.36597674 1.300473365   4.666450     1   53    0    1    0
#> 69      69  4.92655148 1.896022630   6.822574     1   61    0    1    0
#> 70      70  8.50123185 1.560888817  10.062121     1   66    0    1    0
#> 71      71  2.09644996 1.193078713   3.289529     1   66    0    1    0
#> 72      72  3.67824483 1.937894981   5.616140     1   61    1    1    1
#> 73      73 14.00702626 0.299168351  20.371623     0   42    0    1    0
#> 74      74 12.81575901 1.490435606  20.256197     0   62    1    0    1
#> 75      75 13.57336144 0.732833170  14.483309     0   53    1    1    0
#> 76      76 11.10688793 1.886916388  12.993804     1   65    0    0    1
#> 77      77 13.65067103 0.655523577  18.081300     0   57    0    0    1
#> 78      78 12.71901840 1.587176209  15.183450     0   64    1    1    0
#> 79      79  4.15625220 1.393965260   5.550217     1   58    1    1    1
#> 80      80 13.28628838 1.019906234  23.154971     0   61    0    1    0
#> 81      81 12.96133651 1.344858098  36.578636     0   61    1    0    1
#> 82      82 12.74166548 1.564529136  16.565988     0   78    1    0    0
#> 83      83 13.71994020 0.586254409  15.912818     0   72    0    0    0
#> 84      84  9.35946924 0.375226433   9.734696     1   63    0    0    0
#> 85      85  3.01751638 1.844514892   4.862031     1   53    0    0    1
#> 86      86 13.01897227 1.287222342  23.352188     0   75    1    1    0
#> 87      87  8.62617477 1.341378423   9.967553     1   51    0    0    1
#> 88      88 12.86706587 1.439128740  21.357161     0   53    0    0    1
#> 89      89 12.87804488 1.428149728  16.888720     0   57    1    0    1
#> 90      90 11.01227336 1.535467605  12.547741     1   55    1    1    1
#> 91      91  6.08868565 1.881745142   7.970431     1   56    0    1    1
#> 92      92 13.04343279 1.262761817  18.356970     0   78    0    0    0
#> 93      93  6.12439409 1.053863459   7.178258     1   64    1    1    1
#> 94      94 12.96875082 1.337443796  14.688302     0   58    1    0    1
#> 95      95  7.74485260 1.011529981   8.756383     1   72    0    1    0
#> 96      96  3.78752871 1.476114000   5.263643     1   67    0    1    0
#> 97      97  9.29108586 0.510322453   9.801408     1   53    0    1    1
#> 98      98 13.52633582 0.779858791  22.261231     0   59    0    0    1
#> 99      99 10.82174912 1.258490415  12.080240     1   70    0    1    0
#> 100    100 12.37330713 1.932887477  64.793189     0   85    0    0    1
#> 101    101 12.54265890 1.763535714  21.988571     0   81    1    1    0
#> 102    102  0.32258255 1.975701769   2.298284     1   73    1    0    0
#> 103    103 12.27606151 1.458822752  13.734884     1   65    1    0    1
#> 104    104  7.43379836 1.242947810   8.676746     1   48    0    1    1
#> 105    105  4.10562201 1.628172625   5.733795     1   58    1    1    0
#> 106    106 13.02918631 1.277008304  16.655844     0   63    0    0    1
#> 107    107  1.20065513 0.082068380   1.282724     0   70    0    1    1
#> 108    108 12.44408288 1.862111733  16.601945     0   66    1    1    0
#> 109    109  0.87432668 0.316807468   1.191134     1   66    1    0    1
#> 110    110 12.73957392 1.566620693  33.263981     0   83    0    0    1
#> 111    111  4.22494864 1.517477012   5.742426     1   69    0    0    1
#> 112    112 12.59449656 1.711698053  19.436503     0   62    0    1    1
#> 113    113 13.15270437 1.153490246  19.173838     0   61    0    1    1
#> 114    114  3.95788320 1.799326662   5.757210     1   66    0    1    1
#> 115    115  6.00577017 1.786235661   7.792006     1   57    0    0    1
#> 116    116 12.42536031 1.880834302  14.607246     0   79    1    0    1
#> 117    117  8.14734538 1.692034146   9.839380     1   62    0    0    1
#> 118    118  9.64790719 1.520082211  11.167989     1   60    0    0    1
#> 119    119  2.01433411 1.060036130   3.074370     1   56    1    1    0
#> 120    120 12.57623427 0.082493735  12.658728     1   74    0    1    0
#> 121    121 12.49467829 1.811516325  31.410834     0   52    1    1    1
#> 122    122 12.96434451 1.341850099  19.787382     0   60    0    1    1
#> 123    123  5.45453170 1.417705139   6.872237     1   68    1    1    1
#> 124    124  2.50525420 0.267581866   2.772836     1   64    0    0    0
#> 125    125 10.79575829 1.966477632  12.762236     1   68    0    0    0
#> 126    126  2.31264350 1.852535335   4.165179     1   52    0    1    0
#> 127    127  9.48060040 1.570349637  11.050950     1   61    0    0    0
#> 128    128 11.13443937 1.395754128  12.530194     1   48    0    0    1
#> 129    129  3.47616704 0.605083683   4.081251     1   57    1    1    1
#> 130    130  8.13038758 1.690406155   9.820794     1   76    0    1    0
#> 131    131 12.33190613 1.974288478  15.909843     0   68    1    0    0
#> 132    132  8.22418119 1.931379354  10.155561     1   66    0    1    1
#> 133    133  4.85276585 1.548232521   6.400998     1   56    0    1    1
#> 134    134  0.05804212 1.330962458   1.389005     1   59    0    1    1
#> 135    135 12.94488076 1.361313854  17.709736     0   55    1    0    0
#> 136    136  6.02151730 1.299805567   7.321323     1   66    1    1    0
#> 137    137  4.40986751 1.853798855   6.263666     1   73    0    1    1
#> 138    138  8.66874462 0.035897000   8.704642     1   68    0    1    0
#> 139    139  2.00086094 1.877608366   3.878469     1   59    0    0    1
#> 140    140  2.90340600 1.965097878   4.868504     1   71    0    1    0
#> 141    141  1.85636615 0.106096841   1.962463     1   69    0    0    0
#> 142    142 12.79481388 1.511380730  14.306195     1   73    1    0    0
#> 143    143  8.56627347 0.890482330   9.456756     1   61    0    1    0
#> 144    144  6.60371347 0.910377387   7.514091     0   53    1    0    0
#> 145    145 13.28317358 1.023021036  29.058641     0   64    1    1    0
#> 146    146 12.75580602 1.550388596  17.047014     0   68    0    0    0
#> 147    147 11.97807985 0.681479206  12.659559     1   81    1    0    1
#> 148    148  7.94659194 0.721856450   8.668448     1   64    1    0    0
#> 149    149  7.24514393 1.648348195   8.893492     1   69    0    1    0
#> 150    150  5.52341669 1.307768784   6.831185     1   64    0    1    0
#>     analysis_time
#> 1        13.51138
#> 2        13.51138
#> 3        13.51138
#> 4        13.51138
#> 5        13.51138
#> 6        13.51138
#> 7        13.51138
#> 8        13.51138
#> 9        13.51138
#> 10       13.51138
#> 11       13.51138
#> 12       13.51138
#> 13       13.51138
#> 14       13.51138
#> 15       13.51138
#> 16       13.51138
#> 17       13.51138
#> 18       13.51138
#> 19       13.51138
#> 20       13.51138
#> 21       13.51138
#> 22       13.51138
#> 23       13.51138
#> 24       13.51138
#> 25       13.51138
#> 26       13.51138
#> 27       13.51138
#> 28       13.51138
#> 29       13.51138
#> 30       13.51138
#> 31       13.51138
#> 32       13.51138
#> 33       13.51138
#> 34       13.51138
#> 35       13.51138
#> 36       13.51138
#> 37       13.51138
#> 38       13.51138
#> 39       13.51138
#> 40       13.51138
#> 41       13.51138
#> 42       13.51138
#> 43       13.51138
#> 44       13.51138
#> 45       13.51138
#> 46       13.51138
#> 47       13.51138
#> 48       13.51138
#> 49       13.51138
#> 50       13.51138
#> 51       13.51138
#> 52       13.51138
#> 53       13.51138
#> 54       13.51138
#> 55       13.51138
#> 56       13.51138
#> 57       13.51138
#> 58       13.51138
#> 59       13.51138
#> 60       13.51138
#> 61       13.51138
#> 62       13.51138
#> 63       13.51138
#> 64       13.51138
#> 65       13.51138
#> 66       13.51138
#> 67       13.51138
#> 68       13.51138
#> 69       13.51138
#> 70       13.51138
#> 71       13.51138
#> 72       13.51138
#> 73       13.51138
#> 74       13.51138
#> 75       13.51138
#> 76       13.51138
#> 77       13.51138
#> 78       13.51138
#> 79       13.51138
#> 80       13.51138
#> 81       13.51138
#> 82       13.51138
#> 83       13.51138
#> 84       13.51138
#> 85       13.51138
#> 86       13.51138
#> 87       13.51138
#> 88       13.51138
#> 89       13.51138
#> 90       13.51138
#> 91       13.51138
#> 92       13.51138
#> 93       13.51138
#> 94       13.51138
#> 95       13.51138
#> 96       13.51138
#> 97       13.51138
#> 98       13.51138
#> 99       13.51138
#> 100      13.51138
#> 101      13.51138
#> 102      13.51138
#> 103      13.51138
#> 104      13.51138
#> 105      13.51138
#> 106      13.51138
#> 107      13.51138
#> 108      13.51138
#> 109      13.51138
#> 110      13.51138
#> 111      13.51138
#> 112      13.51138
#> 113      13.51138
#> 114      13.51138
#> 115      13.51138
#> 116      13.51138
#> 117      13.51138
#> 118      13.51138
#> 119      13.51138
#> 120      13.51138
#> 121      13.51138
#> 122      13.51138
#> 123      13.51138
#> 124      13.51138
#> 125      13.51138
#> 126      13.51138
#> 127      13.51138
#> 128      13.51138
#> 129      13.51138
#> 130      13.51138
#> 131      13.51138
#> 132      13.51138
#> 133      13.51138
#> 134      13.51138
#> 135      13.51138
#> 136      13.51138
#> 137      13.51138
#> 138      13.51138
#> 139      13.51138
#> 140      13.51138
#> 141      13.51138
#> 142      13.51138
#> 143      13.51138
#> 144      13.51138
#> 145      13.51138
#> 146      13.51138
#> 147      13.51138
#> 148      13.51138
#> 149      13.51138
#> 150      13.51138

# Or use both (whichever happens first)
ex_tte_df |> mutate(
  analysis_time = calc_study_duration(study_time = total_time, observed_time = y,
                                      event_indicator = event,
                                      target_events = 30, target_follow_up = 12)
)
#>     subjid           y    enr_time total_time event cov1 cov2 cov3 cov4
#> 1        1  4.94132575 0.914830838   5.856157     0   62    1    0    0
#> 2        2 13.48784696 0.818347654  34.120433     0   68    1    0    0
#> 3        3  4.08557603 1.234845352   5.320421     1   54    1    0    1
#> 4        4 12.54853421 1.757660399  18.259145     0   37    1    0    0
#> 5        5 11.60144136 1.519444841  13.120886     1   64    0    1    1
#> 6        6 13.06419095 1.242003662  17.094510     0   58    0    1    1
#> 7        7  9.81314580 1.058963733  10.872110     1   67    0    0    1
#> 8        8  8.32546909 1.739833746  10.065303     1   57    0    0    0
#> 9        9 12.22300778 0.005328632  12.228336     1   73    0    0    0
#> 10      10  2.18070722 1.209103289   3.389811     1   65    0    0    0
#> 11      11  1.87453932 0.938280275   2.812820     1   71    0    1    1
#> 12      12 13.03537379 1.270820822  28.926599     0   55    0    0    0
#> 13      13 13.58968582 0.120243722  13.709930     1   63    0    1    0
#> 14      14  3.85157607 0.879423183   4.730999     1   68    0    0    0
#> 15      15  1.23871879 0.028882144   1.267601     1   64    0    1    1
#> 16      16  8.02534813 0.225453192   8.250801     1   82    1    0    0
#> 17      17  4.14391880 0.848946745   4.992866     1   45    0    0    0
#> 18      18 11.65283579 1.893318857  13.546155     0   73    0    0    0
#> 19      19  4.60972545 1.143453675   5.753179     1   65    0    0    1
#> 20      20  6.97459156 0.552640099   7.527232     1   74    0    1    1
#> 21      21 12.78742955 1.518765057  16.883155     0   64    1    0    0
#> 22      22  3.40359707 1.135579805   4.539177     1   38    0    1    0
#> 23      23 10.70343388 1.801794218  12.505228     1   69    1    0    1
#> 24      24  2.57956045 1.801857366   4.381418     1   63    0    0    0
#> 25      25  8.97807945 1.667375555  10.645455     1   55    0    0    1
#> 26      26  1.07397237 0.979580768   2.053553     1   58    0    1    0
#> 27      27 13.45647204 0.849722569  39.847723     0   66    1    1    0
#> 28      28 10.56432229 0.530742681  11.095065     1   90    0    0    1
#> 29      29 13.60629872 0.699895892  14.749382     0   67    0    0    1
#> 30      30  9.66509057 1.287793164  10.952884     1   55    1    0    0
#> 31      31 12.99654787 1.309646739  21.987621     0   84    0    1    0
#> 32      32  0.64072655 0.803063575   1.443790     1   61    1    0    0
#> 33      33 12.41001105 1.896183561  38.443438     0   57    0    0    1
#> 34      34  1.52626737 1.675651089   3.201918     0   53    0    1    1
#> 35      35  5.86099717 0.856978191   6.717975     1   58    0    0    1
#> 36      36 12.54889557 1.757299039  35.511939     0   84    0    0    0
#> 37      37  6.62993839 0.182591691   6.812530     0   74    0    1    0
#> 38      38  9.63003796 0.658219346  10.288257     1   72    1    0    0
#> 39      39 10.85368481 0.882857819  11.736543     1   66    0    1    1
#> 40      40  8.72428218 1.356662716  10.080945     1   68    1    1    1
#> 41      41 12.49996325 1.806231361  33.485484     0   71    0    0    1
#> 42      42  1.80517492 0.790120253   2.595295     1   67    0    0    1
#> 43      43  1.11816335 0.322014023   1.440177     1   67    1    0    0
#> 44      44  9.64377430 1.638116490  11.281891     1   75    1    0    1
#> 45      45  3.69218383 1.943357178   5.635541     1   63    0    0    0
#> 46      46 10.17686106 1.896202200  12.073063     1   68    1    1    0
#> 47      47 12.78212492 1.524069688  16.497252     0   56    1    0    0
#> 48      48 13.32399282 0.982201790  23.498352     0   81    0    0    0
#> 49      49  2.35230418 0.775603405   3.127908     1   79    1    1    1
#> 50      50 11.79858863 1.343483454  13.142072     1   47    0    0    1
#> 51      51 13.10875140 1.197443212  14.972792     0   63    0    0    0
#> 52      52 13.74045847 0.565736139  35.274691     0   70    1    1    0
#> 53      53  9.08213968 1.864018058  10.946158     1   70    0    0    0
#> 54      54  9.89000393 0.615678150  10.505682     1   73    0    1    0
#> 55      55 13.74927445 0.556920157  16.984721     0   61    1    0    1
#> 56      56 11.57290297 0.207753573  11.780657     1   53    0    1    0
#> 57      57  0.82052430 0.859835137   1.680359     1   68    0    0    1
#> 58      58 13.78174289 0.524451723  14.406547     0   50    1    1    0
#> 59      59  6.14890044 1.806120872   7.955021     1   71    0    0    1
#> 60      60  6.51968350 1.235754918   7.755438     1   72    0    0    0
#> 61      61 12.46898554 1.837209076  18.695948     0   46    0    0    1
#> 62      62 13.05387397 1.252320643  15.624370     0   74    0    0    0
#> 63      63  8.08749313 1.101818035   9.189311     1   63    0    1    1
#> 64      64  7.78842293 1.889389239   9.677812     0   58    0    0    1
#> 65      65  2.25040964 1.499576018   3.749986     1   69    0    1    0
#> 66      66 12.69776464 1.608429969  17.494943     0   78    1    1    0
#> 67      67  9.16383335 1.168298671  10.332132     1   81    1    0    1
#> 68      68  3.36597674 1.300473365   4.666450     1   53    0    1    0
#> 69      69  4.92655148 1.896022630   6.822574     1   61    0    1    0
#> 70      70  8.50123185 1.560888817  10.062121     1   66    0    1    0
#> 71      71  2.09644996 1.193078713   3.289529     1   66    0    1    0
#> 72      72  3.67824483 1.937894981   5.616140     1   61    1    1    1
#> 73      73 14.00702626 0.299168351  20.371623     0   42    0    1    0
#> 74      74 12.81575901 1.490435606  20.256197     0   62    1    0    1
#> 75      75 13.57336144 0.732833170  14.483309     0   53    1    1    0
#> 76      76 11.10688793 1.886916388  12.993804     1   65    0    0    1
#> 77      77 13.65067103 0.655523577  18.081300     0   57    0    0    1
#> 78      78 12.71901840 1.587176209  15.183450     0   64    1    1    0
#> 79      79  4.15625220 1.393965260   5.550217     1   58    1    1    1
#> 80      80 13.28628838 1.019906234  23.154971     0   61    0    1    0
#> 81      81 12.96133651 1.344858098  36.578636     0   61    1    0    1
#> 82      82 12.74166548 1.564529136  16.565988     0   78    1    0    0
#> 83      83 13.71994020 0.586254409  15.912818     0   72    0    0    0
#> 84      84  9.35946924 0.375226433   9.734696     1   63    0    0    0
#> 85      85  3.01751638 1.844514892   4.862031     1   53    0    0    1
#> 86      86 13.01897227 1.287222342  23.352188     0   75    1    1    0
#> 87      87  8.62617477 1.341378423   9.967553     1   51    0    0    1
#> 88      88 12.86706587 1.439128740  21.357161     0   53    0    0    1
#> 89      89 12.87804488 1.428149728  16.888720     0   57    1    0    1
#> 90      90 11.01227336 1.535467605  12.547741     1   55    1    1    1
#> 91      91  6.08868565 1.881745142   7.970431     1   56    0    1    1
#> 92      92 13.04343279 1.262761817  18.356970     0   78    0    0    0
#> 93      93  6.12439409 1.053863459   7.178258     1   64    1    1    1
#> 94      94 12.96875082 1.337443796  14.688302     0   58    1    0    1
#> 95      95  7.74485260 1.011529981   8.756383     1   72    0    1    0
#> 96      96  3.78752871 1.476114000   5.263643     1   67    0    1    0
#> 97      97  9.29108586 0.510322453   9.801408     1   53    0    1    1
#> 98      98 13.52633582 0.779858791  22.261231     0   59    0    0    1
#> 99      99 10.82174912 1.258490415  12.080240     1   70    0    1    0
#> 100    100 12.37330713 1.932887477  64.793189     0   85    0    0    1
#> 101    101 12.54265890 1.763535714  21.988571     0   81    1    1    0
#> 102    102  0.32258255 1.975701769   2.298284     1   73    1    0    0
#> 103    103 12.27606151 1.458822752  13.734884     1   65    1    0    1
#> 104    104  7.43379836 1.242947810   8.676746     1   48    0    1    1
#> 105    105  4.10562201 1.628172625   5.733795     1   58    1    1    0
#> 106    106 13.02918631 1.277008304  16.655844     0   63    0    0    1
#> 107    107  1.20065513 0.082068380   1.282724     0   70    0    1    1
#> 108    108 12.44408288 1.862111733  16.601945     0   66    1    1    0
#> 109    109  0.87432668 0.316807468   1.191134     1   66    1    0    1
#> 110    110 12.73957392 1.566620693  33.263981     0   83    0    0    1
#> 111    111  4.22494864 1.517477012   5.742426     1   69    0    0    1
#> 112    112 12.59449656 1.711698053  19.436503     0   62    0    1    1
#> 113    113 13.15270437 1.153490246  19.173838     0   61    0    1    1
#> 114    114  3.95788320 1.799326662   5.757210     1   66    0    1    1
#> 115    115  6.00577017 1.786235661   7.792006     1   57    0    0    1
#> 116    116 12.42536031 1.880834302  14.607246     0   79    1    0    1
#> 117    117  8.14734538 1.692034146   9.839380     1   62    0    0    1
#> 118    118  9.64790719 1.520082211  11.167989     1   60    0    0    1
#> 119    119  2.01433411 1.060036130   3.074370     1   56    1    1    0
#> 120    120 12.57623427 0.082493735  12.658728     1   74    0    1    0
#> 121    121 12.49467829 1.811516325  31.410834     0   52    1    1    1
#> 122    122 12.96434451 1.341850099  19.787382     0   60    0    1    1
#> 123    123  5.45453170 1.417705139   6.872237     1   68    1    1    1
#> 124    124  2.50525420 0.267581866   2.772836     1   64    0    0    0
#> 125    125 10.79575829 1.966477632  12.762236     1   68    0    0    0
#> 126    126  2.31264350 1.852535335   4.165179     1   52    0    1    0
#> 127    127  9.48060040 1.570349637  11.050950     1   61    0    0    0
#> 128    128 11.13443937 1.395754128  12.530194     1   48    0    0    1
#> 129    129  3.47616704 0.605083683   4.081251     1   57    1    1    1
#> 130    130  8.13038758 1.690406155   9.820794     1   76    0    1    0
#> 131    131 12.33190613 1.974288478  15.909843     0   68    1    0    0
#> 132    132  8.22418119 1.931379354  10.155561     1   66    0    1    1
#> 133    133  4.85276585 1.548232521   6.400998     1   56    0    1    1
#> 134    134  0.05804212 1.330962458   1.389005     1   59    0    1    1
#> 135    135 12.94488076 1.361313854  17.709736     0   55    1    0    0
#> 136    136  6.02151730 1.299805567   7.321323     1   66    1    1    0
#> 137    137  4.40986751 1.853798855   6.263666     1   73    0    1    1
#> 138    138  8.66874462 0.035897000   8.704642     1   68    0    1    0
#> 139    139  2.00086094 1.877608366   3.878469     1   59    0    0    1
#> 140    140  2.90340600 1.965097878   4.868504     1   71    0    1    0
#> 141    141  1.85636615 0.106096841   1.962463     1   69    0    0    0
#> 142    142 12.79481388 1.511380730  14.306195     1   73    1    0    0
#> 143    143  8.56627347 0.890482330   9.456756     1   61    0    1    0
#> 144    144  6.60371347 0.910377387   7.514091     0   53    1    0    0
#> 145    145 13.28317358 1.023021036  29.058641     0   64    1    1    0
#> 146    146 12.75580602 1.550388596  17.047014     0   68    0    0    0
#> 147    147 11.97807985 0.681479206  12.659559     1   81    1    0    1
#> 148    148  7.94659194 0.721856450   8.668448     1   64    1    0    0
#> 149    149  7.24514393 1.648348195   8.893492     1   69    0    1    0
#> 150    150  5.52341669 1.307768784   6.831185     1   64    0    1    0
#>     analysis_time
#> 1        5.550217
#> 2        5.550217
#> 3        5.550217
#> 4        5.550217
#> 5        5.550217
#> 6        5.550217
#> 7        5.550217
#> 8        5.550217
#> 9        5.550217
#> 10       5.550217
#> 11       5.550217
#> 12       5.550217
#> 13       5.550217
#> 14       5.550217
#> 15       5.550217
#> 16       5.550217
#> 17       5.550217
#> 18       5.550217
#> 19       5.550217
#> 20       5.550217
#> 21       5.550217
#> 22       5.550217
#> 23       5.550217
#> 24       5.550217
#> 25       5.550217
#> 26       5.550217
#> 27       5.550217
#> 28       5.550217
#> 29       5.550217
#> 30       5.550217
#> 31       5.550217
#> 32       5.550217
#> 33       5.550217
#> 34       5.550217
#> 35       5.550217
#> 36       5.550217
#> 37       5.550217
#> 38       5.550217
#> 39       5.550217
#> 40       5.550217
#> 41       5.550217
#> 42       5.550217
#> 43       5.550217
#> 44       5.550217
#> 45       5.550217
#> 46       5.550217
#> 47       5.550217
#> 48       5.550217
#> 49       5.550217
#> 50       5.550217
#> 51       5.550217
#> 52       5.550217
#> 53       5.550217
#> 54       5.550217
#> 55       5.550217
#> 56       5.550217
#> 57       5.550217
#> 58       5.550217
#> 59       5.550217
#> 60       5.550217
#> 61       5.550217
#> 62       5.550217
#> 63       5.550217
#> 64       5.550217
#> 65       5.550217
#> 66       5.550217
#> 67       5.550217
#> 68       5.550217
#> 69       5.550217
#> 70       5.550217
#> 71       5.550217
#> 72       5.550217
#> 73       5.550217
#> 74       5.550217
#> 75       5.550217
#> 76       5.550217
#> 77       5.550217
#> 78       5.550217
#> 79       5.550217
#> 80       5.550217
#> 81       5.550217
#> 82       5.550217
#> 83       5.550217
#> 84       5.550217
#> 85       5.550217
#> 86       5.550217
#> 87       5.550217
#> 88       5.550217
#> 89       5.550217
#> 90       5.550217
#> 91       5.550217
#> 92       5.550217
#> 93       5.550217
#> 94       5.550217
#> 95       5.550217
#> 96       5.550217
#> 97       5.550217
#> 98       5.550217
#> 99       5.550217
#> 100      5.550217
#> 101      5.550217
#> 102      5.550217
#> 103      5.550217
#> 104      5.550217
#> 105      5.550217
#> 106      5.550217
#> 107      5.550217
#> 108      5.550217
#> 109      5.550217
#> 110      5.550217
#> 111      5.550217
#> 112      5.550217
#> 113      5.550217
#> 114      5.550217
#> 115      5.550217
#> 116      5.550217
#> 117      5.550217
#> 118      5.550217
#> 119      5.550217
#> 120      5.550217
#> 121      5.550217
#> 122      5.550217
#> 123      5.550217
#> 124      5.550217
#> 125      5.550217
#> 126      5.550217
#> 127      5.550217
#> 128      5.550217
#> 129      5.550217
#> 130      5.550217
#> 131      5.550217
#> 132      5.550217
#> 133      5.550217
#> 134      5.550217
#> 135      5.550217
#> 136      5.550217
#> 137      5.550217
#> 138      5.550217
#> 139      5.550217
#> 140      5.550217
#> 141      5.550217
#> 142      5.550217
#> 143      5.550217
#> 144      5.550217
#> 145      5.550217
#> 146      5.550217
#> 147      5.550217
#> 148      5.550217
#> 149      5.550217
#> 150      5.550217