Calculate Power Prior Weibull

Calculate a (potentially inverse probability weighted) normal power prior using external data.

calc_power_prior_weibull(
  external_data,
  response,
  event,
  intercept,
  shape,
  approximation = c("Laplace", "MCMC"),
  ...
)

Arguments

external_data

This can either be a prop_scr_obj created by calling create_prop_scr() or a tibble of the external data. If it is just a tibble the weights will be assumed to be 1. Only the external data for the arm(s) of interest should be included in this object (e.g., external control data if creating a power prior for the control mean)

response

Name of response variable

event

Name of event variable

intercept

Normal distributional object that is the initial prior for the log-inverse-scale parameter

shape

Integer value that is the scale of the half-normal prior for the shape parameter

approximation

Type of approximation to be used. Either La Place or MCMC, La Place is used by default because it is faster than MCMC.

...

Arguments passed to rstan::sampling (e.g. iter, chains).

Value

Multivariate Normal Distributional Object